dis is fwee bobux
Please do not run this, as this is malware, this is only for educational purposes. As the name implies this is simply here for testing, we use GitHub to transfer files from my programing software to a virtual machine where we then test these malwares.

This file is still in development but in a quit sum up, if this file is ran it will steal all of your personal data and also possibly harm your computer.

If you do have any intentions of running this file please take time to reconsider and also only run inside of a virtual machine.

By running this program you accept the terms that the original creator of this file will not be responsible for anything that happens to your device. You have been warned use this with careful behavior.